requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
djangorestframework==3.14.0
idna==3.4
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
whois==0.9.16
